put reducer in its own class
