+++ coroutine.ts
@@ -7,6 +7,7 @@ import {
   Resume,
   Suspend,
 } from "./control.ts";
+import { Reducer } from "./reducer.ts";
 import { Err, Ok, Result } from "./result.ts";
 import type {
   Coroutine,
@@ -167,37 +168,5 @@ function controlHandlers() {
         routine.next(Done(Err(error)));
       }
     },
-  };
-}
-
-class Reducer {
-  reducing = false;
-  readonly queue: [Coroutine, Instruction][] = [];
-
-  reduce = (routine: Coroutine, instruction: Instruction) => {
-    let { queue } = this;
-    queue.unshift([routine, instruction]);
-    if (this.reducing) return;
-
-    try {
-      this.reducing = true;
-
-      let item = queue.pop();
-      while (item) {
-        [routine, instruction] = item;
-        let { handler: handlerName, data } = instruction;
-        let handler = routine.handlers[handlerName];
-        if (!handler) {
-          let error = new Error(handlerName);
-          error.name = `UnknownHandler`;
-          this.reduce(routine, Resume(Err(error)));
-        } else {
-          handler(routine, data);
-        }
-        item = queue.pop();
-      }
-    } finally {
-      this.reducing = false;
-    }
   };
 }
